♻️ FIX: 일기 설정 페이지 추상화 작업
